Remove some unused code
