+++ appservice/package.json
@@ -59,6 +59,6 @@
         "vscode": "^1.1.5"
     },
     "engines": {
-        "vscode": "^1.28.0"
+        "vscode": "^1.23.0"
     }
 }
